Reorganize logging configurations and update dependencies
This commit involves a significant reorganization of the logging configurations. The previous logging configurations located in the application/commons were removed, and new ones were added under application/logs-config.

Corresponding updates were made in a number of build.gradle files to reflect this structural change. In addition, the logger setup in the MapperTest.java file was tidied, a README.md file was added to application/logs-config, and a line was added to .gitignore to ignore the logs folder.