Add dependency management plugin and implement pagination
The spring-boot and spring-cloud build.gradle files use the io.spring.dependency-management plugin to help with dependency management. Also, a new function "limitPage" was added to the BookService interface, its implementation, and the BookMapper interface to add pagination support in the app. Configuration has been added for MyBatis Plus Interceptor for Postgres database and logging level is added to the application.yml file. Lombok version is removed in dependencies for version consistency.