Add dynamic select functionality and corresponding tests
This commit introduces a dynamic select feature that permits query parameterization in student and teacher entities. It involves creating new tables in the database, adding new entities for their corresponding representation and also providing unit tests for the feature. For the functionality, 'DynamicMapper' is used for the querying and 'GroupQueryParam' for parameter setting.