🗑️ Removed entire camel netty-api module
The camel netty-api module which comprised of functionalities for starting/stopping a TCP server, and their corresponding http APIs had been entirely removed from the project. This involved deleting all related classes, files and removing its reference from application settings. The reasons behind this significant change are not specified but it's important to take note of this impacts.